Relay modified timestamp back to client for each update
For flownotes to function properly when accessed from different machines with
potentially out-of-sync clocks, all timestamps must be issued by the server.
Relay them back to the client for each change, so that the client can always
the correct time. This also serves as a confirmation that the data was
succesfully stored by the server.