Add a button to delete a note
Notes are not hard-deleted but are marked for deletion in the database. This
enables propagation of the deletion to parallel sessions and serves as a
recycle bin for inadvertent deletions. Eventually we might want to create a
cleanup routine for the database that runs an actual SQL DELETE on notes that
where marked for deletion before some time threshold.