Support adding/appending to notes from the command line
This feature also detects when the input is an email message
(judging by the input starting with "From ") and is able to parse
the content out of any valid MIME message. To facilitate this we
need a new dependency for a MIME parser. This adds 3 MB of PHP code
to the project, but it is only loaded when it is actually needed, so
the impact is limited.